git-svn-id: http://citadel.physiology.northwestern.edu/svn/s1_analysis@1320 984b1e3c-1c57-0410-ae62-c18f90d3b645

+++ proc/Tucker/PDs/PD_plot.m
@@ -71,7 +71,7 @@ if (include_unsorted && length(u1)~=length(moddepth))
     u1=u1(~u1(:,2),:);
 end
 %set_outputs
-outdata=[u1(:,1),pds,moddepth,errs];
+outdata=[double(u1(:,1)),pds,moddepth,errs];
 
 %% identify channels with excessive modulation depth
 %% identify channels with excessive modulation depth
